# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/households/{householdId}/recipes/{recipeId}" : {
      "get" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "recipeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "recipeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRecipe"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "recipeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/admin/assistant/agents/{agentId}" : {
      "put" : {
        "tags" : [ "assistant-admin-controller" ],
        "operationId" : "update_3",
        "parameters" : [ {
          "name" : "agentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AgentConfigUpdate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentConfigDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "assistant-admin-controller" ],
        "operationId" : "reset",
        "parameters" : [ {
          "name" : "agentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/households/{householdId}/recipes/import" : {
      "post" : {
        "tags" : [ "recipe-import-controller" ],
        "operationId" : "upload",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportSessionInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/households/{householdId}/recipes/import/{importId}/confirm" : {
      "post" : {
        "tags" : [ "recipe-import-controller" ],
        "operationId" : "confirm",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "importId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmImport"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportOutcome"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/households/{householdId}/recipes/import/url" : {
      "post" : {
        "tags" : [ "recipe-import-controller" ],
        "operationId" : "fromUrl",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UrlImport"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportSessionInfo"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/assistant/agents/{agentId}/test" : {
      "post" : {
        "tags" : [ "assistant-admin-controller" ],
        "operationId" : "test",
        "parameters" : [ {
          "name" : "agentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentTestResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/households/{householdId}/assistant/agents" : {
      "get" : {
        "tags" : [ "agent-controller" ],
        "operationId" : "list_6",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AgentDefinition"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/assistant/providers" : {
      "get" : {
        "tags" : [ "assistant-admin-controller" ],
        "operationId" : "providers_1",
        "parameters" : [ {
          "name" : "refresh",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProviderDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/households/{householdId}/recipes/import/{importId}" : {
      "delete" : {
        "tags" : [ "recipe-import-controller" ],
        "operationId" : "discard",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "importId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "CreateRecipe" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "servings" : {
            "type" : "integer",
            "format" : "int32"
          },
          "notes" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "steps" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "ingredients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IngredientDto"
            }
          },
          "imageMimeType" : {
            "type" : "string"
          },
          "imageData" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
      "IngredientDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "number",
            "format" : "double"
          },
          "unit" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          }
        }
      },
      "RecipeDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "householdId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "servings" : {
            "type" : "integer",
            "format" : "int32"
          },
          "notes" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "steps" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "imageMimeType" : {
            "type" : "string"
          },
          "imageData" : {
            "type" : "string"
          },
          "ingredients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IngredientDto"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "AgentConfigUpdate" : {
        "type" : "object",
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          },
          "provider" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "temperature" : {
            "type" : "number",
            "format" : "double"
          },
          "maxTokens" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "AgentConfigDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "nameKey" : {
            "type" : "string"
          },
          "descriptionKey" : {
            "type" : "string"
          },
          "supportsMedia" : {
            "type" : "boolean"
          },
          "orchestrator" : {
            "type" : "boolean"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "provider" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "temperature" : {
            "type" : "number",
            "format" : "double"
          },
          "maxTokens" : {
            "type" : "integer",
            "format" : "int32"
          },
          "defaults" : {
            "$ref" : "#/components/schemas/ModelDefaults"
          },
          "overridden" : {
            "type" : "boolean"
          },
          "available" : {
            "type" : "boolean"
          }
        }
      },
      "ModelDefaults" : {
        "type" : "object",
        "properties" : {
          "provider" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "temperature" : {
            "type" : "number",
            "format" : "double"
          },
          "maxTokens" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "CandidateInfo" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "snippet" : {
            "type" : "string"
          },
          "structured" : {
            "type" : "boolean"
          }
        }
      },
      "ImportSessionInfo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "source" : {
            "type" : "string"
          },
          "candidates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CandidateInfo"
            }
          },
          "aiAvailable" : {
            "type" : "boolean"
          }
        }
      },
      "ConfirmImport" : {
        "type" : "object",
        "properties" : {
          "candidates" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "minItems" : 1
          }
        },
        "required" : [ "candidates" ]
      },
      "ImportFailure" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
      "ImportOutcome" : {
        "type" : "object",
        "properties" : {
          "imported" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RecipeDto"
            }
          },
          "failures" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImportFailure"
            }
          }
        }
      },
      "UrlImport" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "url" ]
      },
...
      "AgentTestResult" : {
        "type" : "object",
        "properties" : {
          "ok" : {
            "type" : "boolean"
          },
          "latencyMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "text" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          }
        }
      },
...
      "AgentDefinition" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "nameKey" : {
            "type" : "string"
          },
          "descriptionKey" : {
            "type" : "string"
          },
          "supportsMedia" : {
            "type" : "boolean"
          },
          "toolNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "toolDescription" : {
            "type" : "string"
          },
          "orchestrator" : {
            "type" : "boolean"
          },
          "model" : {
            "type" : "string"
          }
        }
      },
...
      "ProviderDto" : {
        "type" : "object",
        "properties" : {
          "provider" : {
            "type" : "string"
          },
          "available" : {
            "type" : "boolean"
          },
          "models" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "error" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}